Updated cmds so can save author

+++ commands/bot/deletesnippet.js
@@ -1,12 +1,11 @@
-const Discord = require('discord.js');
-const { botColour } = require('../../config.json');
+const { owner } = require('../../config.json');
 const fs = require('fs');
 
 module.exports = {
 	name: 'deletesnippet',
-	description: 'Info on the dev',
-	category: 'Owner',
-	execute(message, client, args) {
+	description: 'Delete your snippet',
+	category: 'Bot',
+	async execute(message, client, args) {
 
 		if(!args[0]) return message.channel.send('You havent defined a valid file name! Example of running the command: `!deletesnippet dictionary txt`');
 
@@ -14,7 +13,13 @@ module.exports = {
 
 		if (!fs.existsSync(`snippets/${filename}`)) return message.channel.send('A snippet with this name doesnt exists!');
 
-		const snippet = args.join(' ');
+		const fileDB = args[0].replace('.', '-');
+
+		const authorID = await client.db.get(fileDB);
+
+		const author = client.users.cache.get(authorID);
+
+		if(message.author.id !== authorID && message.author.id !== owner) return message.channel.send('You didnt create this snippet!');
 
 		try {
 			fs.unlinkSync(`snippets/${filename}`);
